Implement authorization on requests

+++ Services/CharacterService.cs
@@ -1,43 +1,48 @@
 using AutoMapper;
 using Microsoft.EntityFrameworkCore;
+using System.Security.Claims;
+using Microsoft.AspNetCore.Identity;
 
 public class CharacterService : ICharacterService
 {
     private readonly DataContext _context;
     private readonly IMapper _mapper;
+    private readonly IHttpContextAccessor _httpContextAccessor;
 
-    public CharacterService(IMapper mapper, DataContext context)
+    public CharacterService(IMapper mapper, DataContext context, IHttpContextAccessor httpContextAccessor)
     {
         _mapper = mapper;
         _context = context;
+        _httpContextAccessor = httpContextAccessor;
     }
 
-    public async Task<ServiceResponse<List<GetCharacterDTO>>> GetCharacters(int userId) 
+    public async Task<ServiceResponse<List<GetCharacterDTO>>> GetCharacters()
     {
         ServiceResponse<List<GetCharacterDTO>> serviceResponse = new ServiceResponse<List<GetCharacterDTO>>();
-        List<Character> dbCharacters = await _context.Characters.Where(c => c.User.Id == userId).ToListAsync();
+        List<Character> dbCharacters = await _context.Characters.Where(c => c.User.Id == GetUserId()).ToListAsync();
         serviceResponse.Data = (dbCharacters.Select(c => _mapper.Map<GetCharacterDTO>(c))).ToList();
-        
+
         return serviceResponse;
     }
 
     public async Task<ServiceResponse<GetCharacterDTO>> GetCharacter(int id)
     {
         ServiceResponse<GetCharacterDTO> serviceResponse = new ServiceResponse<GetCharacterDTO>();
-        Character dbCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id);
+        Character dbCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
         serviceResponse.Data = _mapper.Map<GetCharacterDTO>(dbCharacter);
 
         return serviceResponse;
     }
 
-    public async Task<ServiceResponse<List<GetCharacterDTO>>> AddCharacter(AddCharacterDTO newCharacter) 
+    public async Task<ServiceResponse<List<GetCharacterDTO>>> AddCharacter(AddCharacterDTO newCharacter)
     {
         ServiceResponse<List<GetCharacterDTO>> serviceResponse = new ServiceResponse<List<GetCharacterDTO>>();
         Character character = _mapper.Map<Character>(newCharacter);
+        character.User = await _context.Users.FirstOrDefaultAsync(u => u.Id == GetUserId());
 
         await _context.Characters.AddAsync(character);
         await _context.SaveChangesAsync();
-        serviceResponse.Data = (_context.Characters.Select(c => _mapper.Map<GetCharacterDTO>(c))).ToList();
+        serviceResponse.Data = (_context.Characters.Where(c => c.User.Id == GetUserId()).Select(c => _mapper.Map<GetCharacterDTO>(c))).ToList();
 
         return serviceResponse;
     }
@@ -47,39 +52,56 @@ public class CharacterService : ICharacterService
         ServiceResponse<GetCharacterDTO> serviceResponse = new ServiceResponse<GetCharacterDTO>();
         try
         {
-            Character character = await _context.Characters.FirstOrDefaultAsync(c => c.Id == updatedCharacter.Id);
-            character.Name = updatedCharacter.Name;
-            character.Class = updatedCharacter.Class;
-            character.Defense = updatedCharacter.Defense;
-            character.HitPoints = updatedCharacter.HitPoints;
-            character.Intelligence = updatedCharacter.Intelligence;
-            character.Strength = updatedCharacter.Strength;
+            Character character = await _context.Characters.Include(c => c.User).FirstOrDefaultAsync(c => c.Id == updatedCharacter.Id);
+            if (character.User.Id == GetUserId())
+            {
+                character.Name = updatedCharacter.Name;
+                character.Class = updatedCharacter.Class;
+                character.Defense = updatedCharacter.Defense;
+                character.HitPoints = updatedCharacter.HitPoints;
+                character.Intelligence = updatedCharacter.Intelligence;
+                character.Strength = updatedCharacter.Strength;
 
-            _context.Characters.Update(character);
-            await _context.SaveChangesAsync();
+                _context.Characters.Update(character);
+                await _context.SaveChangesAsync();
 
-            serviceResponse.Data = _mapper.Map<GetCharacterDTO>(character);
-        } catch (Exception ex)
+                serviceResponse.Data = _mapper.Map<GetCharacterDTO>(character);
+            }
+            else
+            {
+                serviceResponse.Success = false;
+                serviceResponse.Message = "Character not found";
+            }
+        }
+        catch (Exception ex)
         {
             serviceResponse.Success = false;
             serviceResponse.Message = ex.Message;
         }
-        
+
         return serviceResponse;
     }
 
     public async Task<ServiceResponse<List<GetCharacterDTO>>> DeleteCharacter(int id)
     {
         ServiceResponse<List<GetCharacterDTO>> serviceResponse = new ServiceResponse<List<GetCharacterDTO>>();
-        try 
+        try
         {
-            Character character = await _context.Characters.FirstAsync(c => c.Id == id);
-            _context.Characters.Remove(character);
+            Character character = await _context.Characters.FirstAsync(c => c.Id == id && c.User.Id == GetUserId());
+            if (character != null)
+            {
+                _context.Characters.Remove(character);
+                await _context.SaveChangesAsync();
+                serviceResponse.Data = (_context.Characters.Where(c => c.User.Id == GetUserId()).Select(c => _mapper.Map<GetCharacterDTO>(c))).ToList();
+            }
+            else
+            {
+                serviceResponse.Success = false;
+                serviceResponse.Message = "Character not found";
+            }
 
-            await _context.SaveChangesAsync();
-
-            serviceResponse.Data = (_context.Characters.Select(c => _mapper.Map<GetCharacterDTO>(c))).ToList();
-        } catch (Exception ex) 
+        }
+        catch (Exception ex)
         {
             serviceResponse.Success = false;
             serviceResponse.Message = ex.Message;
@@ -87,4 +109,6 @@ public class CharacterService : ICharacterService
 
         return serviceResponse;
     }
+
+    private int GetUserId() => int.Parse(_httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier));
 }
